Use simulationID in output csv file and RNG
Add simulationID to the iter ID in the output csv file.  Use the RNG seed value from the file together with the simulationID from command line to seed the RNG.  Make the third command line parameter and rngSeed model parameter compulsory (to simplify things), and set a default fixed RNG seed for use in tests (so that all the tests don't have to explicitly seed the RNG).