Reseed RNG before each test
Add a new 'SimulationTest' base class for tests, which loads the default
parameters and reseeds the RNG before tests run.  (This is not done for the
tests where it doesn't make sense, such as tests which don't use the RNG or
default parameters.)  Previously, the results of tests were different depending
upon where they were run - although the RNG was being seeded with a fixed value,
it was not being reseeded between every test run.