Output DailyStats via a map
This way we don't lose track of the header names or accidentally put the
wrong value to the wrong column. Much easier to extend with more
metrics.

Note: The R code will need to reflect the new column names (keeping
these as the field names is much easier since we can just do a little
meta-programming to define the map)