Use example_params for testing
This avoids needing to maintain 2 parameters files.

We also remove some of the testing for parameter
input (ParameterIOTest), replacing this with the "isValid"
checks that are already there (I don't think it's particularly necessary
to test the gson is doing what it's meant to do for specific values,
only that we can catch errors/parse successfully).

We still have some extra files but these are for specific
invalid/default argument cases.