Add a commented-out regression test
This adds a commented-out assert to testBaseLine() to check that results have
not changed. The idea is that this assert should be uncommented while making
changes (such as refactoring) which should not affect the results, but
generally kept commented out, so that we don't get lots of git merge conflicts.

For this to work, temporary changes to these commented out lines should not
be committed to git.