Refactor DailyStats
Change DailyStats to have a lot of public IntValue fields (where IntValue is
a new Class), with these fields added to relevant lists for calculated totals
(e.g. of daily deaths or infections) as they're declared.  The order that the
fields are created is the same as the order in the csv file.