Fix integer overflow bug
When adding transport contacts (in network generator mode), the number of pairs
passed to nextBinomial was a 32bit integer, but with larger populations it needs
to be a 64bit integer. Since we are using an RNG which we can't easily change,
this quick fix adds a wrapper to rng.nextBinomial() which makes an approximation
for large (>32bit) numbers of pairs.